+++ packages/conatus_schedule/test/schedule_decode_test.dart
@@ -1,4 +1,5 @@
 import 'package:conatus_foundation/conatus_foundation.dart';
+import 'package:conatus_schedule/conatus_schedule.dart';
 import 'package:test/test.dart';
 
 SessionEvent changeEvent(int seq, Object? data) => SessionEvent.create(