+++ packages/conatus_schedule/test/schedule_delivery_test.dart
@@ -2,6 +2,7 @@ import 'package:conatus_agent/conatus_agent.dart';
 import 'package:conatus_core/conatus_core.dart';
 import 'package:conatus_foundation/conatus_foundation.dart';
 import 'package:conatus_llm/conatus_llm.dart';
+import 'package:conatus_schedule/conatus_schedule.dart';
 import 'package:test/test.dart';
 
 class _ScriptedProvider implements LlmProvider {