+++ packages/conatus_schedule/test/schedule_time_test.dart
@@ -1,4 +1,4 @@
-import 'package:conatus_foundation/conatus_foundation.dart';
+import 'package:conatus_schedule/conatus_schedule.dart';
 import 'package:test/test.dart';
 
 Matcher inputError(String code) => throwsA(isA<ScheduleInputException>()